Integrated existsInTable function into Authors model. Cleaned up Authors model code.

+++ api/authors/delete.php
@@ -5,18 +5,26 @@ $author = new Author($db);
 // Get raw posted data
 $data = json_decode(file_get_contents("php://input"));
 
-// Set id to update
-$author->id = $data->id;
-
-// Delete author
-if($author->delete()) {
-  echo json_encode(
-  array(
-    'id' => $author->id
-  ));
-} else {
+// Check for required params
+if (isset($data->id)) {
+  // Set id to delete
+  $author->id = $data->id;
+  
+  // Delete author
+  if($author->delete()) {
+    echo json_encode(
+    array(
+      'id' => $author->id
+    ));
+  } else {
     echo json_encode(
     array(
       'message' => 'author_id Not Found'
     ));
+  }
+} else {
+  echo json_encode(
+  array(
+    'message' => 'Missing Required Parameters'
+  ));
 }
